src/requirements.txt
click==7.1.2
tqdm==4.50.2
pystardog==0.9.9
